[feat] add programmatic interface to create-svelte
