+++ .changeset/fifty-foxes-tan.md
@@ -1,0 +1,5 @@
+---
+'create-svelte': patch
+---
+
+Add a programmatic interface to create-svelte
