+++ packages/create-svelte/bin.js
@@ -1,10 +1,10 @@
 #!/usr/bin/env node
 import fs from 'fs';
 import path from 'path';
-import { fileURLToPath } from 'url';
 import { bold, cyan, gray, green, red } from 'kleur/colors';
 import prompts from 'prompts';
-import { mkdirp, copy } from './utils.js';
+import { create } from './index.js';
+import { dist } from './utils.js';
 
 // prettier-ignore
 const disclaimer = `
@@ -36,8 +36,6 @@ async function main() {
 				process.exit(1);
 			}
 		}
-	} else {
-		mkdirp(cwd);
 	}
 
 	const options = /** @type {import('./types/internal').Options} */ (
@@ -83,10 +81,7 @@ async function main() {
 		])
 	);
 
-	const name = path.basename(path.resolve(cwd));
-
-	write_template_files(options.template, options.typescript, name, cwd);
-	write_common_files(cwd, options, name);
+	await create(cwd, options);
 
 	console.log(bold(green('\nYour project is ready!')));
 
@@ -124,133 +119,6 @@ async function main() {
 
 	console.log(`\nTo close the dev server, hit ${bold(cyan('Ctrl-C'))}`);
 	console.log(`\nStuck? Visit us at ${cyan('https://svelte.dev/chat')}\n`);
-}
-
-/**
- * @param {string} template
- * @param {boolean} typescript
- * @param {string} name
- * @param {string} cwd
- */
-function write_template_files(template, typescript, name, cwd) {
-	const dir = dist(`templates/${template}`);
-	copy(`${dir}/assets`, cwd, (name) => name.replace('DOT-', '.'));
-	copy(`${dir}/package.json`, `${cwd}/package.json`);
-
-	const manifest = `${dir}/files.${typescript ? 'ts' : 'js'}.json`;
-	const files = /** @type {import('./types/internal').File[]} */ (
-		JSON.parse(fs.readFileSync(manifest, 'utf-8'))
-	);
-
-	files.forEach((file) => {
-		const dest = path.join(cwd, file.name);
-		mkdirp(path.dirname(dest));
-
-		fs.writeFileSync(dest, file.contents.replace(/~TODO~/g, name));
-	});
-}
-
-/**
- *
- * @param {string} cwd
- * @param {import('./types/internal').Options} options
- * @param {string} name
- */
-function write_common_files(cwd, options, name) {
-	const shared = dist('shared.json');
-	const { files } = /** @type {import('./types/internal').Common} */ (
-		JSON.parse(fs.readFileSync(shared, 'utf-8'))
-	);
-
-	const pkg_file = path.join(cwd, 'package.json');
-	const pkg = /** @type {any} */ (JSON.parse(fs.readFileSync(pkg_file, 'utf-8')));
-
-	files.forEach((file) => {
-		const include = file.include.every((condition) => matchesCondition(condition, options));
-		const exclude = file.exclude.some((condition) => matchesCondition(condition, options));
-
-		if (exclude || !include) return;
-
-		if (file.name === 'package.json') {
-			const new_pkg = JSON.parse(file.contents);
-			merge(pkg, new_pkg);
-		} else {
-			const dest = path.join(cwd, file.name);
-			mkdirp(path.dirname(dest));
-			fs.writeFileSync(dest, file.contents);
-		}
-	});
-
-	pkg.dependencies = sort_keys(pkg.dependencies);
-	pkg.devDependencies = sort_keys(pkg.devDependencies);
-	pkg.name = toValidPackageName(name);
-
-	fs.writeFileSync(pkg_file, JSON.stringify(pkg, null, '  '));
-}
-
-/**
- * @param {import('./types/internal').Condition} condition
- * @param {import('./types/internal').Options} options
- * @returns {boolean}
- */
-function matchesCondition(condition, options) {
-	return condition === 'default' || condition === 'skeleton'
-		? options.template === condition
-		: options[condition];
-}
-
-/**
- * @param {any} target
- * @param {any} source
- */
-function merge(target, source) {
-	for (const key in source) {
-		if (key in target) {
-			const target_value = target[key];
-			const source_value = source[key];
-
-			if (
-				typeof source_value !== typeof target_value ||
-				Array.isArray(source_value) !== Array.isArray(target_value)
-			) {
-				throw new Error('Mismatched values');
-			}
-
-			merge(target_value, source_value);
-		} else {
-			target[key] = source[key];
-		}
-	}
-}
-
-/** @param {Record<string, any>} obj */
-function sort_keys(obj) {
-	if (!obj) return;
-
-	/** @type {Record<string, any>} */
-	const sorted = {};
-	Object.keys(obj)
-		.sort()
-		.forEach((key) => {
-			sorted[key] = obj[key];
-		});
-
-	return sorted;
-}
-
-/** @param {string} path */
-function dist(path) {
-	return fileURLToPath(new URL(`./dist/${path}`, import.meta.url).href);
-}
-
-/** @param {string} name */
-function toValidPackageName(name) {
-	return name
-		.trim()
-		.toLowerCase()
-		.replace(/\s+/g, '-')
-		.replace(/^[._]/, '')
-		.replace(/[^a-z0-9~.-]+/g, '-');
 }
 
 main();
